refactor(task-board): split policy and dispatch modules
Keep both files under the source-length cap after the WP3 additions: move the policy unit tests into a companion policy_tests.rs and extract the spawn-agent decision logic (subject enrichment, fail-closed switches, kill switch, graph and built-in gate selection) into dispatch_spawn_policy.rs. Also add the additive PolicyApprovalGrant wire type describing the durable grant shape for the resolution surface.
